python/agents/RAG/pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = "Answer questions related to Vertex AI documentation using Vertex AI RAG Engine"
authors = [
    { name = "Hangfei Lin", email = "hangfei@google.com" },
    { name = "Chouaieb Nemri", email = "nemri@google.com" },
]
license = { text = "Apache License 2.0" }
readme = "README.md"
requires-python = ">=3.11,<3.13"

# --- CORE DEPENDENCIES ---
dependencies = [
    "google-cloud-aiplatform[adk,agent-engines]>=1.108.0",
    "google-adk>=1.10.0",
    "pydantic-settings>=2.8.1",
    "tabulate>=0.9.0",
    "google-auth>=2.36.0",
    "requests>=2.32.3",
    "llama-index>=0.12",
    "openinference-instrumentation-google-adk>=0.1.0",
    "openinference-instrumentation>=0.1.34",
    "arize-otel>=0.8.2",
    "arize>=7.36.0",
    "ruff>=0.4.6",
    "mypy>=1.15.0",
]

[project.optional-dependencies]
lint = [
    "ruff>=0.4.6",
    "mypy>=1.15.0",
    "codespell>=2.2.0",
    "types-pyyaml>=6.0.12.20240917",
    "types-requests>=2.32.0.20240914",
]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "pytest-mock>=3.14.0",
    "scikit-learn>=1.6.1",
    "pytest-cov>=6.0.0",
    "pytest-asyncio>=0.26.0",
    "google-adk[eval]>=1.10.0",
    "nest-asyncio>=1.6.0",
    "agent-starter-pack>=0.32.0",
    "ruff>=0.4.6",
    "mypy>=1.15.0",
    "types-requests>=2.32.0.20240914",
]

[tool.agent-starter-pack]
example_question = "What are the key business segments mentioned in Alphabet's 2024 10-K report?"

[tool.agent-starter-pack.settings]
agent_directory = "rag"

# --- CI/CD AND TESTING METADATA ---
[tool.agent-starter-pack.testing]
RAG_CORPUS = "projects/adk-devops/locations/us-central1/ragCorpora/456"
ENV = "test"

[tool.ruff]
extend = "../../../pyproject.toml"

[tool.ruff.lint]
extend-select = ["I"] 
ignore = ["E501", "C901"]

[tool.ruff.lint.per-file-ignores]
"**/__init__.py" = ["E402", "F401", "I001"]

[tool.ruff.lint.isort]
known-first-party = ["rag"]

[tool.mypy]
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
no_implicit_optional = true
check_untyped_defs = true
disallow_subclassing_any = true
warn_incomplete_stub = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_unreachable = true
follow_imports = "silent"
ignore_missing_imports = true
explicit_package_bases = true
disable_error_code = ["misc", "no-untyped-call", "no-any-return"]
exclude = [".venv"]

[tool.codespell]
ignore-words-list = "rouge"
skip = "./locust_env/*,uv.lock,.venv,./frontend,**/*.ipynb"

[tool.pytest.ini_options]
pythonpath = "."
asyncio_default_fixture_loop_scope = "function"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"